+++ src/bookstore/src/main/java/com/bookstore/models/Author.java
@@ -12,27 +12,20 @@ import javax.persistence.OneToMany;
 import com.fasterxml.jackson.annotation.JsonManagedReference;
 
 @Entity
-public class Author {
+public class Author extends Person {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Integer id;
 
-    private String name;
-    private int age;
     private String nationaly;
 
     @JsonManagedReference
     @OneToMany(cascade = CascadeType.ALL, mappedBy = "author")
     private List<Book> books;
 
-    public Author() {
-    }
-
     public Author(String name, int age, String nationaly) {
-        super();
-        setName(name);
-        setAge(age);
+        super(name, age);
         setNationaly(nationaly);
     }
 
@@ -43,22 +36,6 @@ public class Author {
 
     public void setNationaly(String nationaly) {
         this.nationaly = nationaly;
-    }
-
-    public void setAge(int age) {
-        this.age = age;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public int getAge() {
-        return age;
     }
 
     public String getNationaly() {
