+++ src/bookstore/src/main/java/com/bookstore/models/Employee.java
@@ -1,0 +1,38 @@
+package com.bookstore.models;
+
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+
+@Entity
+public class Employee extends Person {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Integer id;
+
+    private String address;
+
+    public Employee(String name, int age, String address) {
+        super(name, age);
+        setAddress(address);
+    }
+
+    public void setId (Integer id)
+    {
+        this.id = id;
+    }
+
+    public void setAddress(String address) {
+        this.address = address;
+    }
+
+    public String getAddress() {
+        return address;
+    }
+
+    public Integer getId() {
+        return id;
+    }
+}
