+++ src/bookstore/src/main/java/com/bookstore/models/Person.java
@@ -1,0 +1,29 @@
+package com.bookstore.models;
+
+
+public class Person {
+
+    private String name;
+    private int age;
+
+    public Person(String name, int age) {
+        setName(name);
+        setAge(age);
+    }
+
+    public void setAge(int age) {
+        this.age = age;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public int getAge() {
+        return age;
+    }
+}
